+++ Config.lua
@@ -89,21 +89,3 @@ function Config:Initialize()
 end
 
 function Config:Clone(tbl)
-    local copy = {}
-    for k, v in pairs(tbl) do
-        copy[k] = type(v) == "table" and self:Clone(v) or v
-    end
-    return copy
-end
-
-function Config:Get(key)
-    -- If using character settings and key exists in character DB, use it
-    if ZenHUDDB.useCharacterSettings and ZenHUDCharDB[key] ~= nil then
-        return ZenHUDCharDB[key]
-    end
-    -- Otherwise use account-wide setting
-    return ZenHUDDB[key]
-end
-
-function Config:Set(key, value)
-    -- Set to appropriate DB based on mode